Add TOS site and notif servers to app config
