+++ src/components/scenes/TermsOfServiceScene.js
@@ -3,9 +3,8 @@
 import * as React from 'react'
 import { WebView } from 'react-native-webview'
 
+import { config } from '../../theme/appConfig.js'
 import { SceneWrapper } from '../common/SceneWrapper.js'
-
-const WEB_URI = 'https://edge.app/tos/'
 
 type Props = {}
 
@@ -13,7 +12,7 @@ export class TermsOfServiceComponent extends React.Component<Props> {
   render() {
     return (
       <SceneWrapper background="body" hasTabs={false}>
-        <WebView source={{ uri: WEB_URI }} />
+        <WebView source={{ uri: config.termsOfServiceSite }} />
       </SceneWrapper>
     )
   }
