+++ src/types/types.js
@@ -351,11 +351,13 @@ export type AppConfig = {
   lightTheme: Theme,
   backgroundImageServers?: string[],
   referralServers?: string[],
+  notificationServers: string[],
   supportsEdgeLogin: boolean,
   knowledgeBase: string,
   supportSite: string,
   phoneNumber: string,
   website: string,
+  termsOfServiceSite: string,
   appStore: string
 }
